move binaries and libraries to a subdirectory of /opt
