+++ images/cfw/etc/init.d/S01sd_watchdog
@@ -2,7 +2,7 @@
 
 start() {
 	printf "Starting sd_watchdog"
-	start-stop-daemon -S -m -b -p /var/run/sd_watchdog.pid --exec /opt/sd_watchdog
+	start-stop-daemon -S -m -b -p /var/run/sd_watchdog.pid --exec /opt/x1plus/bin/sd_watchdog
 }
 stop() {
 	printf "Stopping sd_watchdog"
