+++ images/cfw/etc/init.d/S72x1plusd
@@ -4,7 +4,7 @@
 
 start() {
 	printf "Starting x1plusd: "
-	start-stop-daemon -S -m -b -p /var/run/x1plusd.pid --exec /opt/python/bin/python3 -- /opt/x1plusd.py
+	start-stop-daemon -S -m -b -p /var/run/x1plusd.pid --exec /opt/python/bin/python3 -- -m x1plus.services.x1plusd
     [ $? = 0 ] && echo "OK" || echo "FAIL"
 }
 stop() {
