+++ images/cfw/etc/init.d/S99perf_log
@@ -5,7 +5,7 @@
 start() {
 	if x1p_get_setting boot.perf_log perf_log; then
 		printf "Starting perf_log"
-		start-stop-daemon -S -m -b -p /var/run/perf_log.pid --exec /opt/perf_log.sh
+		start-stop-daemon -S -m -b -p /var/run/perf_log.pid --exec /opt/x1plus/bin/perf_log.sh
 	else
 		printf "perf_log disabled. Skipping start."
 	fi
