+++ images/cfw/etc/profile.d/x1plus.sh
@@ -1,1 +1,1 @@
-export PATH=$PATH:/opt/python3/bin
+export PATH=/opt/python3/bin:/opt/x1plus/bin:$PATH
