--- images/cfw/opt/bbl_screen_patch
@@ -1,2 +1,0 @@
-#!/bin/sh
-LD_PRELOAD=/opt/printer_ui.so exec /usr/bin/bbl_screen "$@"
