--- images/cfw/opt/dds.py
@@ -1,1 +1,0 @@
-../../../installer/dds.py
