--- images/cfw/opt/evdev_fix.so
@@ -1,1 +1,0 @@
-../../../bbl_screen-patch/evdev_fix.so
