--- images/cfw/opt/libdds_intf.so
@@ -1,1 +1,0 @@
-../../../prebuilt/libdds_intf.so
