--- images/cfw/opt/uncurl.so
@@ -1,1 +1,0 @@
-../../../bbl_screen-patch/uncurl.so
