--- images/cfw/opt/workaround_ifaddrs.so
@@ -1,1 +1,0 @@
-../../../bbl_screen-patch/workaround_ifaddrs.so
