+++ images/cfw/opt/x1plus/bin/bbl_screen_patch
@@ -1,0 +1,2 @@
+#!/bin/sh
+LD_PRELOAD=/opt/x1plus/lib/printer_ui.so exec /usr/bin/bbl_screen "$@"
