+++ images/cfw/opt/x1plus/bin/perf_log.sh
