+++ images/cfw/opt/x1plus/lib/evdev_fix.so
@@ -1,0 +1,1 @@
+../../../../../bbl_screen-patch/evdev_fix.so
