+++ images/cfw/opt/x1plus/lib/python/x1plus/dds.py
@@ -1,0 +1,1 @@
+../../../../../../../installer/dds.py
