+++ images/cfw/opt/x1plus/lib/python/x1plus/libdds_intf.so
@@ -1,0 +1,1 @@
+../../../../../../../prebuilt/libdds_intf.so
