+++ images/cfw/opt/x1plus/lib/uncurl.so
@@ -1,0 +1,1 @@
+../../../../../bbl_screen-patch/uncurl.so
