+++ images/cfw/opt/x1plus/lib/workaround_ifaddrs.so
@@ -1,0 +1,1 @@
+../../../../../bbl_screen-patch/workaround_ifaddrs.so
